community/livekit_opentelemetry/pyproject.toml
[project]
name = "agent_telephony_livekit_pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "livekit-agents>=0.12.15",
    "livekit-plugins-deepgram>=0.6.19",
    "livekit-plugins-elevenlabs>=0.7.13",
    "livekit-plugins-openai>=0.11.0",
    "livekit-plugins-silero>=0.7.4",
    "livekit-plugins-turn-detector>=0.4.2",
    "opentelemetry-api>=1.30.0",
    "opentelemetry-exporter-gcp-monitoring>=1.9.0a0",
    "opentelemetry-sdk>=1.30.0",
    "python-dotenv==1.0.1",
    "restack-ai>=0.0.80",
]

[tool.hatch.build.targets.sdist]
include = ["src"]

[tool.hatch.build.targets.wheel]
include = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
